+++ common/RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # dxCommon
 
+## in develop
+...
+
 ## 0.11.5 (2023-03-07)
 
 * Implemented immutable linked list to parse instance names