+++ common/RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # dxCommon
 
-## in develop
+## unreleased
 ...
 
 ## 0.11.5 (2023-03-07)