+++ common/RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # dxCommon
 
-## unreleased
+## Unreleased
 ...
 
 ## 0.11.5 (2023-03-07)